chore: Rename namespace-select-filter component file
