+++ packages/core/src/renderer/components/namespace-select-filter/component.tsx
